Weseeker/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weseeker"
version = "0.1.0"
description = "Windows 本地智能文件管家 Agent"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2",
    "langgraph>=1.1",
    "langchain-mcp-adapters>=0.2",
    "langchain-openai>=0.4",
    "langchain-text-splitters>=0.3",
    "langchain-experimental>=0.3",
    "langchain-chroma>=0.2",
    "mcp>=1.9",
    "langgraph-checkpoint-sqlite>=2.0",
    "python-pptx>=1.0",
    "python-docx>=1.1",
    "PyMuPDF>=1.24",
    "openpyxl>=3.1",
    "Pillow>=10.0",
    "chromadb>=0.5",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic-settings>=2.5",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[project.scripts]
weseeker-agent = "listeners.cli:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
